Fail sooner if Android device is offline
Interestingly, `adb connect` returns a successful exit code if it times out trying to connect to a device over Tailscale. This PR adds an additional check that we actually connected to a device so the job will fail then if the device is offline instead of later as part of trying to use the device.